refactor: extract resource registry into @vm0/core
Move the pure resource-registry data and find/list helpers from the CLI
into `@vm0/core/resource-registry` as the single source of truth. The CLI
keeps a thin re-export shim so all existing `./resource-registry` imports
are unchanged.

This is an inert refactor with no behavior change; it unblocks PR2/PR3,
which need the registry server-side to validate generationStyle ids.
